refactor post

+++ client/src/hooks/post/useDeletePost.ts
@@ -2,12 +2,19 @@ import apiClient from "@/utils/apiClient";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { toast } from "sonner";
 import { postKeys } from "./postKeys";
+import { deletePostSchema } from "@/validation/postsSchema";
 
 interface DeletePostResponse {
   message: string;
 }
 const deletePost = async (postId: string): Promise<DeletePostResponse> => {
-  const url = `/api/posts/${postId}`;
+  const parsed = deletePostSchema.safeParse({ postId });
+
+  if (!parsed.success) {
+    throw new Error(`${parsed.error.issues[0].message} - ${parsed.error.issues[0].path}`);
+  }
+
+  const url = `/api/posts/${parsed.data.postId}`;
   const res = await apiClient.delete(url);
   return res.data;
 };
